Use snap to install VS Code when possible

+++ install_dependencies.sh
@@ -58,17 +58,22 @@ fi
 echo
 # Only install VS code if not installed yet and not WSL
 if ! command -v code > /dev/null && ! grep -q Microsoft /proc/version > /dev/null; then
-  echo Install VS code
+  if command -v snap > /dev/null; then
+    echo Install VS code using snap
+    sudo snap install code --classic
+  else
+    echo Install VS code via apt-get
+    curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
+    sudo install -o root -g root -m 644 packages.microsoft.gpg /usr/share/keyrings/
+    sudo sh -c 'echo "deb [arch=amd64 signed-by=/usr/share/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
 
-  curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
-  sudo install -o root -g root -m 644 packages.microsoft.gpg /usr/share/keyrings/
-  sudo sh -c 'echo "deb [arch=amd64 signed-by=/usr/share/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
+    sudo apt-get install apt-transport-https
+    sudo apt-get update
+    sudo apt-get install code # or code-insiders
 
-  sudo apt-get install apt-transport-https
-  sudo apt-get update
-  sudo apt-get install code # or code-insiders
+    rm packages.microsoft.gpg
+  fi
 
-  rm packages.microsoft.gpg
 else
   echo VS Code already installed
 fi
